Move config docs to own readme; update main readme
Config docs moved to own readme because Snakedeploy will only copy
contents of the `config` folder but not the main readme, therefore the
docs should be in the same subfolder as the config files themselves.

Briefly describe pipeline in the main readme and link to homepages of
dependencies.

Update install instructions, but Snakedeploy still not fully tested so
that subsection is incomplete.