Handle case where multilvl method not specified in config
If `multilvl` classification method not specified in config file,
Snakemake nonetheless parses the rule which looks for threshold values
set in the config file. This leads to a KeyError if user has removed
those fields from the config file because they are unused.

Substitute these with None values instead if `multilvl` key is not
present udner `class_thresholds` in the config file.